Split groovy-lint.js
