+++ src/codenarc-factory.js
@@ -12,7 +12,7 @@ const { evaluateRange, evaluateVariables, getSourceLines } = require("./utils.js
 // Build codenarc options //
 ////////////////////////////
 
-const notCodeNarcRuleNames = ["Groovy", "Jenkins", "all"];
+const notCodeNarcRuleNames = ["Groovy", "Jenkinsfile", "all"];
 
 // Convert NPM-groovy-lint into codeNarc arguments
 // Create temporary files if necessary
@@ -189,6 +189,7 @@ async function parseCodeNarcResult(options, codeNarcBaseDir, tmpXmlFileName) {
         }
     }
     // Complete with files with no error
+
     result.files = files;
     return result;
 }
@@ -210,7 +211,9 @@ async function manageCreateRuleSetFile(options) {
         // Create groovy ruleset definition
         let ruleSetSource = `ruleset {\n\n    description 'Generated by npm-groovy-lint (https://github.com/nvuillam/npm-groovy-lint#readme)'\n\n`;
         for (const rule of ruleSetsDef) {
-            ruleSetSource += `    ${rule.name}\n`;
+            if (!(npmGroovyLintRules[rule.name] && npmGroovyLintRules[rule.name].isCodeNarcRule === false)) {
+                ruleSetSource += `    ${rule.name}\n`;
+            }
         }
         ruleSetSource += `\n}\n`;
         // Write file
