+++ test/lint-build.test.js
@@ -14,8 +14,12 @@ describe('Lint with executables (jdeploy-bundle)', () => {
             '--loglevel', 'warning',
             '--verbose'
         ];
-        const { stdout } = await exec('npm-groovy-lint ' + params.join(' '));
-        assert(stdout && stdout.includes('warning'), 'Script failure');
+        const { stdout, stderr } = await exec('npm-groovy-lint ' + params.join(' '));
+        if (stderr) {
+            console.error(stderr);
+        }
+        assert(stdout, 'stdout is set');
+        assert(stdout.includes('warning'), 'stdout should contain word "warning"');
     });
     it('(EXE:file) should generate json console output', async () => {
         const params = [
@@ -23,8 +27,12 @@ describe('Lint with executables (jdeploy-bundle)', () => {
             '--rulesets', '"Groovy"',
             '--output', 'json'
         ];
-        const { stdout } = await exec('npm-groovy-lint ' + params.join(' '));
-        assert(stdout && stdout.includes('"totalFilesWithErrorsNumber"'), 'Script failure');
+        const { stdout, stderr } = await exec('npm-groovy-lint ' + params.join(' '));
+        if (stderr) {
+            console.error(stderr);
+        }
+        assert(stdout, 'stdout is set');
+        assert(stdout.includes('"totalFilesWithErrorsNumber"'), 'stdout should contain word "totalFilesWithErrorsNumber"');
     });
 
     it('(EXE:file) should generate codenarc HTML file report', async () => {
@@ -36,7 +44,7 @@ describe('Lint with executables (jdeploy-bundle)', () => {
             '-maxPriority1Violations=0',
             '-report="html:ReportTestCodenarc.html"'];
         await exec('npm-groovy-lint ' + params.join(' '));
-        assert(fse.existsSync('ReportTestCodenarc.html'), 'Script failure');
+        assert(fse.existsSync('ReportTestCodenarc.html'), 'html CodeNarc report has been generated');
         fse.removeSync('ReportTestCodenarc.html');
     });
 
@@ -49,8 +57,8 @@ describe('Lint with executables (jdeploy-bundle)', () => {
             '-maxPriority1Violations=0',
             '-report="xml:ReportTestCodenarc.xml"'];
         await exec('npm-groovy-lint ' + params.join(' '));
-        assert(fse.existsSync('ReportTestCodenarc.xml'), 'Script failure');
-        //       fse.removeSync('ReportTestCodenarc.xml');
+        assert(fse.existsSync('ReportTestCodenarc.xml'), 'xml CodeNarc report has been generated');
+        fse.removeSync('ReportTestCodenarc.xml');
     });
 
     it('(EXE:file) should run on a Jenkinsfile', async () => {
@@ -58,16 +66,24 @@ describe('Lint with executables (jdeploy-bundle)', () => {
             '--path', ' "jdeploy-bundle/lib/example"',
             '-r', '"Jenkinsfile"',
             '--verbose'];
-        const { stdout } = await exec('npm-groovy-lint ' + params.join(' '));
-        assert(stdout && stdout.includes('warning'), 'Script failure');
+        const { stdout, stderr } = await exec('npm-groovy-lint ' + params.join(' '));
+        if (stderr) {
+            console.error(stderr);
+        }
+        assert(stdout, 'stdout is set');
+        assert(stdout.includes('warning'), 'stdout should contain word "warning"');
     });
 
     it('(EXE:help) should show npm-groovy-lint help', async () => {
         const params = [
             '-h'
         ];
-        const { stdout } = await exec('npm-groovy-lint ' + params.join(' '));
-        assert(stdout.includes('--verbose'), 'Script failure');
+        const { stdout, stderr } = await exec('npm-groovy-lint ' + params.join(' '));
+        if (stderr) {
+            console.error(stderr);
+        }
+        assert(stdout, 'stdout is set');
+        assert(stdout.includes('--verbose'), 'stdout should contain word "--verbose"');
     });
 
     it('(EXE:help) should show codenarc help', async () => {
@@ -75,8 +91,12 @@ describe('Lint with executables (jdeploy-bundle)', () => {
             '--codenarcargs',
             '-help'
         ];
-        const { stdout } = await exec('npm-groovy-lint ' + params.join(' '));
-        assert(stdout.includes('where OPTIONS are zero or more command-line options'), 'Script failure');
+        const { stdout, stderr } = await exec('npm-groovy-lint ' + params.join(' '));
+        if (stderr) {
+            console.error(stderr);
+        }
+        assert(stdout, 'stdout is set');
+        assert(stdout.includes('where OPTIONS are zero or more command-line options'), 'stdout should contain word "where OPTIONS are zero or more command-line options"');
     });
